uart: changed arguments
-int UART1_send(uint8_t *data, uint8_t size);
-uint8_t UART1_ReadData( uint8_t *buff, int size);
+uint8_t UART1_send(uint8_t *data, uint8_t size);
+uint8_t UART1_ReadData( uint8_t *buff, uint8_t size);

+++ Dali/include/r_uart.h
@@ -351,6 +351,6 @@ Macro definitions
 
 void UART1_init();
 void UART1_start();
-int UART1_send(uint8_t *data, uint8_t size);
-uint8_t UART1_ReadData( uint8_t *buff, int size);
+uint8_t UART1_send(uint8_t *data, uint8_t size);
+uint8_t UART1_ReadData( uint8_t *buff, uint8_t size);
 #endif
